+++ lib/src/auth/phone/phone.service.dart
@@ -55,9 +55,10 @@ class PhoneService {
 
   /// This method is invoked when user submit sms code, then it will begin
   /// verification process.
-  Future verifySMSCode({required VoidCallback success, required ErrorCallback error}) {
-    PhoneAuthCredential credential =
-        PhoneAuthProvider.credential(verificationId: verificationId, smsCode: smsCode);
+  Future verifySMSCode(
+      {required VoidCallback success, required ErrorCallback error}) {
+    PhoneAuthCredential credential = PhoneAuthProvider.credential(
+        verificationId: verificationId, smsCode: smsCode);
 
     return verifyCredential(credential, success: success, error: error);
   }
@@ -114,8 +115,13 @@ class PhoneService {
         /// This code is only for Android, and this method is invoked after
         /// automatic sms verification has succeed.
         /// Note that, not all Android phone support automatic sms resolution.
-        verificationCompleted: (c) => verifyCredential(c, success: success, error: error),
-        verificationFailed: error,
+        verificationCompleted: (PhoneAuthCredential c) {
+          verifyCredential(c, success: success, error: error);
+        },
+        verificationFailed: (FirebaseAuthException e) {
+          print(e);
+          error(e);
+        },
         codeSent: (String verificationId, i) {
           this.verificationId = verificationId;
           codeSent(verificationId);