+++ lib/src/user/user.model.dart
@@ -1,0 +1,9 @@
+class UserModel {
+  UserModel({
+    required this.name,
+    required this.photoUrl,
+  });
+
+  String name;
+  String photoUrl;
+}