+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+
+## [0.2.7] - Chat
+
+- 1:1 chat functionality has been added. see README.md for details.
+
 ## [0.2.6] - Phone sign-in code submit progress
 
 - Display progress bar on sms code submit.